Initial commit
Created the project structure and started coding. Set up database, models and an
endpoint to create users. I also set up authentication using JWT. set up
environment variables and schemas to on what data to receive when creating a user
and a token and also what data to send after a user has logged in or created a user.