Created the API endpoints for tasks
Updated the tasks model by adding one more field "is_active: bool" to it and
add the "is_active and id" to the TaskResponseModel in the schemas.py file.
And also created a new file "tasks.py" which contains all the endpoints to the
tasks api. Finally I created a utility function to update all tasks "is_active"
to True or False if the "task.time_due > the time now" or "task.time_due < the time now)
respectively.